refactor: move duplicate code to env_utils.py
